Changed unix to linux in filenames
Most of those files were Linux specific, and in the case where they
were generic *nix, I think it makes more sense to label them as Linux
since it is much more common now.  There is still on instance in User
Agents that is Unix specific, so it still has unix in its name.